+++ src/claspar/utils.py
@@ -5,14 +5,12 @@ table creation function, error handling and io functions.
 
 import json
 import logging
-import os
 from pathlib import Path
 
 import pandas as pd
 import yaml
-from onyx import OnyxClient, OnyxConfig, OnyxEnv
 from onyx_analysis_helper import onyx_analysis_helper_functions as oa
-from profiler import __version__ as pv
+from pandas.core.frame import DataFrame
 
 
 ##############
@@ -47,27 +45,22 @@ class InputError(ClasParError):
 ########
 # Onyx #
 
-# Set up onyx config
-CONFIG = OnyxConfig(
-    domain=os.environ[OnyxEnv.DOMAIN],
-    token=os.environ[OnyxEnv.TOKEN],
-)
 
-
-@oa.call_to_onyx
-def get_input_data(sample_id: str, server: str) -> tuple[list[pd.DataFrame], int]:
+def get_input_data(sample_id: str, server: str) -> tuple[list[pd.DataFrame], list[dict], int]:
     """
     Get the input data from Onyx. Decorated to handle errors suitably.
     :param sample_id: ID of the sample (climb-id).
     :param server: the server to query.
     :return: tuple of dataframes (list) and exitcode (int).
     """
-    with OnyxClient(CONFIG) as client:
-        record = client.get(
-            project=server,
-            climb_id=sample_id,
-            include=["classifier_calls", "alignment_results", "sylph_results"],
-        )
+    fields_to_include = ["classifier_calls", "alignment_results", "sylph_results"]
+    record, onyx_versions, onyx_exitcode = oa.get_data_and_versions_from_onyx(
+        sample_id, server, fields=fields_to_include
+    )
+
+    if onyx_exitcode != 0:
+        logging.error("Issue occured with Onyx query and getting data, see above. Exiting cleanly.")
+        return [pd.DataFrame(), pd.DataFrame(), pd.DataFrame()], [], onyx_exitcode
 
     try:
         alignment_results_df = pd.DataFrame(record["alignment_results"])
@@ -78,7 +71,7 @@ def get_input_data(sample_id: str, server: str) -> tuple[list[pd.DataFrame], int
         logging.error("Could not find key %s in Onyx Record. Exiting cleanly." % (e))  # noqa
         exitcode = 1
 
-    return [alignment_results_df, sylph_results_df, classifier_calls_df], exitcode
+    return [alignment_results_df, sylph_results_df, classifier_calls_df], onyx_versions, exitcode
 
 
 ###################
@@ -90,8 +83,9 @@ def create_analysis_fields(
     domain: str,
     classifier: str,
     record_id: str,
-    thresholds: dict[str, int | str],
-    profile_table_name: str,
+    thresholds_dict: dict[str, int | str],
+    onyx_versions: list,
+    tool_versions: dict,
     headline_result: str,
     results: dict,
     server: str,
@@ -101,8 +95,10 @@ def create_analysis_fields(
     :param domain: str, one of 'bacteria', 'virus', 'fungi' etc
     :param classifier: the type of classifier being reported in the table (kraken or sylph)
     :param record_id: Climb ID for sample
-    :param thresholds: Dictionary containing criteria used to filter
-    :param profile_table_name: Name of the profile tables file used to assign profiles. Acts as versioning.
+    :param thresholds_dict: Dictionary containing criteria used to filter, which gets added to the
+    methods field as 'thresholds': {thresholds_dict}
+    :param onyx_versions: list of versions from onyx - must be from when data was first queried.
+    :param tool_versions: dict of tools, databases, files etc and their versions.
     :param headline_result: Short description of main result
     :param results: Dictionary containing results
     :param server: Server code is running on, one of "server" or "synthscape"
@@ -110,10 +106,6 @@ def create_analysis_fields(
     onyx_analysis: Class containing required fields for input to onyx analysis table.
     exitcode: Exit code for checks - will be 0 if all checks passed, 1 if any checks failed
     """
-    # Add the profiler version to the thresholds dict so it can be loaded into the methods
-    thresholds["profiler_version"] = pv
-    thresholds["profile_tables_version"] = profile_table_name
-
     onyx_analysis = oa.OnyxAnalysis()  # set up class
     # Add analysis details
     onyx_analysis.add_analysis_details(
@@ -124,20 +116,25 @@ def create_analysis_fields(
     # Add metadata about the pipeline/package
     onyx_analysis.add_package_metadata(package_name="claspar")
     # Check that the methods were parsed by the class
-    methods_fail = onyx_analysis.add_methods(methods_dict=thresholds)
+    methods_versions_fail = onyx_analysis.add_versions_to_methods(
+        onyx_versions=onyx_versions, tool_versions=tool_versions
+    )
+
+    # Reformat the thresholds_dict:
+    methods_dict: dict[str, dict[str, int | str]] = {"thresholds": thresholds_dict}
+    # Check that additional methods are parsed by the class
+    methods_fail = onyx_analysis.add_methods(methods_dict)
+
     # Check that the results were parsed by the class
     results_fail = onyx_analysis.add_results(top_result=headline_result, results_dict=results)
+
     # Add info about sample and server (server/synthscape)
     onyx_analysis.add_server_records(sample_id=record_id, server_name=server)
     # Check the final object using the helper method
     required_field_fail, attribute_fail = onyx_analysis.check_analysis_object(publish_analysis=False)
+    # print(onyx_analysis.methods)
     # If any fail, raise exit code.
-    if any(  # noqa: SIM108
-        [methods_fail, results_fail, required_field_fail, attribute_fail]
-    ):  # noqa SIM108
-        exitcode = 1
-    else:
-        exitcode = 0
+    exitcode = 1 if any([methods_fail, methods_versions_fail, results_fail, required_field_fail, attribute_fail]) else 0
 
     return onyx_analysis, exitcode
 
@@ -156,7 +153,7 @@ def read_samplesheet(path_to_samplesheet: Path | str) -> tuple[list[pd.DataFrame
     the sylph results and the classifier calls. Note that any of these could be empty dataframes!
     """
     exitcode = 0
-    samplesheet_df = pd.read_csv(path_to_samplesheet, sep="\t")
+    samplesheet_df: DataFrame = pd.read_csv(path_to_samplesheet, sep="\t")
     try:
         json_str = samplesheet_df["full_Onyx_json"].iloc[0]
     except KeyError as k: